Revert "refactor: migrate to new unified APIs (#1267)"
This reverts commit 9d0e27d7df419d7997a5c07e3759b5eb700dbf8c.
