+++ apps/api/.env.example
@@ -1,2 +1,1 @@
 DATABASE_URL=postgres://postgres:password@localhost:5432/api
-IS_INDEXER=true
