+++ apps/ui/src/helpers/constants.ts
@@ -10,8 +10,8 @@ export const APP_NAME = 'Snapshot';
 
 export const SIDEKICK_URL = 'https://sh5.co';
 
-export const UNIFIED_API_URL = 'https://api.snapshot.box';
-export const UNIFIED_API_TESTNET_URL = 'https://testnet-api.snapshot.box';
+export const UNIFIED_API_URL = 'https://api-u.snapshot.box';
+export const UNIFIED_API_TESTNET_URL = 'https://testnet-api-u.snapshot.box';
 
 export const HELPDESK_URL = 'https://help.snapshot.box';
